backend/requirements.txt
# FastAPI and Related Dependencies
fastapi==0.109.0
uvicorn==0.24.0
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Database Dependencies
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication and Security
python-dotenv==1.0.0
email-validator==2.1.0

# Optional: For type hinting and validation
typing-extensions==4.9.0
